+++ src/game/game_entity.h
@@ -9,93 +9,93 @@ namespace GameLib {
 
 /** @brief possible anchor point types */
 enum AnchorType {
-  ANCHOR_NONE,
-  ANCHOR_CENTER,  /// entity centeres around center of sprite
-  ANCHOR_BASELINE /// entity centers around center-bottom of sprite
+    ANCHOR_NONE,
+    ANCHOR_CENTER,  /// entity centeres around center of sprite
+    ANCHOR_BASELINE /// entity centers around center-bottom of sprite
 };
 
 class GameEntity {
 public:
-  /**
-  * @brief construct
-  */
-  // GameEntity(void);
+    /**
+    * @brief construct
+    */
+    // GameEntity(void);
 
-  /**
-  * @brief construct with ready made physical aspect
-  * @param p ready made physical aspect - subclass responsible for memory!
-  * @param r ready made renderable aspect- subclass responsible for memory!
-  */
-  explicit GameEntity(Physical *p, Renderable *r);
+    /**
+    * @brief construct with ready made physical aspect
+    * @param p ready made physical aspect - subclass responsible for memory!
+    * @param r ready made renderable aspect- subclass responsible for memory!
+    */
+    explicit GameEntity (Physical *p, Renderable *r);
 
-  /**
-   * @brief destruct
-   */
-  virtual ~GameEntity(void);
+    /**
+     * @brief destruct
+     */
+    virtual ~GameEntity (void);
 
-  /**
-   * @brief the main update function for an entity
-   * @param dt time delta
-   */
-  virtual void Update(double dt);
+    /**
+     * @brief the main update function for an entity
+     * @param dt time delta
+     */
+    virtual void Update (float dt);
 
-  /**
-   * @brief manually move the entity
-   * @param dx x delta
-   * @param dy y delta
-   */
-  void Move(double dx, double dy);
+    /**
+     * @brief manually move the entity
+     * @param dx x delta
+     * @param dy y delta
+     */
+    void Move (float dx, float dy);
 
-  /**
-   * @brief set position
-   * @param x x pos
-   * @param y y pos
-   * @param z z pos
-   */
-  void SetPosition(double x, double y, double z = 0);
+    /**
+     * @brief set position
+     * @param x x pos
+     * @param y y pos
+     * @param z z pos
+     */
+    void SetPosition (float x, float y, float z = 0);
 
-  Point GetPosition();
+    Point GetPosition();
 
-  /**
-   * @brief set a uniqueentity id
-   */
-  void SetName(const std::string &name);
+    /**
+     * @brief set a uniqueentity id
+     */
+    void SetName (const std::string &name);
 
-  /**
-   * @brief get a uniqueentity id
-   */
-  std::string GetName();
+    /**
+     * @brief get a uniqueentity id
+     */
+    std::string GetName();
 
-  /**
-   * @brief like a rpc call
-   * @param params list of params
-   */
-  virtual void Call(std::vector<std::string> params);
+    /**
+     * @brief like a rpc call
+     * @param params list of params
+     */
+    virtual void Call (std::vector<std::string> params);
 
-  /// a game entity has a physical and graphical aspect
-  Physical *physical;
-  Renderable *renderable;
+    /// a game entity has a physical and graphical aspect
+    Physical *physical;
+    Renderable *renderable;
 
-  /// easy access without dereferencing each time
-  Vector3 &velocity;
+    /// easy access without dereferencing each time
+    Vector3 &velocity;
 
-  /// center point
-  AnchorType anchor_type;
+    /// center point
+    AnchorType anchor_type;
 
-  /// overlay (hud)
-  bool hud = false;
+    /// overlay (hud)
+    bool hud = false;
 
-  /// keep the list unique
-  static std::set<std::string> entity_names;
+    /// keep the list unique
+    static std::set<std::string> entity_names;
 
 protected:
-  /**
-   * @brief align the sprite accordig to physical position, sprite size and
-   * anchor type
-   */
-  void anchor();
+    /**
+     * @brief align the sprite accordig to physical position, sprite size and
+     * anchor type
+     */
+    void anchor();
 
-  /// unique name
-  std::string name;
+    /// unique name
+    std::string name;
 };
 } // GameLib
