+++ src/graphics/primitives.cpp
@@ -7,61 +7,61 @@
 
 namespace GameLib {
 
-Color Primitives::outline_color(0, 0, 0, 0);
-Color Primitives::fill_color(0, 0, 0, 0);
-Color Primitives::last_outline_color(0, 0, 0, 0);
-Color Primitives::last_fill_color(0, 0, 0, 0);
+Color Primitives::outline_color (0, 0, 0, 0);
+Color Primitives::fill_color (0, 0, 0, 0);
+Color Primitives::last_outline_color (0, 0, 0, 0);
+Color Primitives::last_fill_color (0, 0, 0, 0);
 
 // ------------------------------------------------------------
 // Circle
 // ------------------------------------------------------------
-void Primitives::Circle(Window &target, unsigned int x, unsigned int y, unsigned int radius) {
-    sf::CircleShape circle(radius);
-    circle.setFillColor(sf::Color(fill_color.r, fill_color.g, fill_color.b, fill_color.a));
-    circle.setOutlineThickness(1);
-    circle.setOutlineColor(sf::Color(outline_color.r, outline_color.g, outline_color.b, outline_color.a));
-    circle.setPosition(x - radius, y - radius);
-    target.Draw(circle);
+void Primitives::Circle (Window &target, unsigned int x, unsigned int y, unsigned int radius) {
+    sf::CircleShape circle (radius);
+    circle.setFillColor (sf::Color (fill_color.r, fill_color.g, fill_color.b, fill_color.a));
+    circle.setOutlineThickness (1);
+    circle.setOutlineColor (sf::Color (outline_color.r, outline_color.g, outline_color.b, outline_color.a));
+    circle.setPosition (x - radius, y - radius);
+    target.Draw (circle);
 }
 
 // ------------------------------------------------------------
 // Triangle
 // ------------------------------------------------------------
-void Primitives::Triangle(Window &target, unsigned int x1, unsigned int y1, unsigned int x2, unsigned int y2, unsigned int x3, unsigned int y3) {
+void Primitives::Triangle (Window &target, unsigned int x1, unsigned int y1, unsigned int x2, unsigned int y2, unsigned int x3, unsigned int y3) {
 
     // create an array of 3 vertices that define a triangle primitive
-    sf::VertexArray triangle(sf::Triangles, 3);
+    sf::VertexArray triangle (sf::Triangles, 3);
 
     // define the position of the triangle's points
-    triangle[0].position = sf::Vector2f(x1, y1);
-    triangle[1].position = sf::Vector2f(x2, y2);
-    triangle[2].position = sf::Vector2f(x3, y3);
+    triangle[0].position = sf::Vector2f (x1, y1);
+    triangle[1].position = sf::Vector2f (x2, y2);
+    triangle[2].position = sf::Vector2f (x3, y3);
 
     // define the color of the triangle's points
-    triangle[0].color = sf::Color(fill_color.r, fill_color.g, fill_color.b, fill_color.a);
+    triangle[0].color = sf::Color (fill_color.r, fill_color.g, fill_color.b, fill_color.a);
     triangle[1].color = triangle[0].color;
     triangle[2].color = triangle[0].color;
 
-    target.Draw(triangle);
+    target.Draw (triangle);
 }
 
 // ------------------------------------------------------------
 // Rectangle
 // ------------------------------------------------------------
-void Primitives::Rectangle(Window &target, unsigned int x, unsigned int y, unsigned int w, unsigned int h) {
+void Primitives::Rectangle (Window &target, unsigned int x, unsigned int y, unsigned int w, unsigned int h) {
     sf::RectangleShape rectangle;
-    rectangle.setSize(sf::Vector2f(w, h));
-    rectangle.setPosition(sf::Vector2f(x, y));
-    rectangle.setFillColor(sf::Color(fill_color.r, fill_color.g, fill_color.b, fill_color.a));
-    rectangle.setOutlineThickness(1);
-    rectangle.setOutlineColor(sf::Color(outline_color.r, outline_color.g, outline_color.b, outline_color.a));
-    target.Draw(rectangle);
+    rectangle.setSize (sf::Vector2f (w, h));
+    rectangle.setPosition (sf::Vector2f (x, y));
+    rectangle.setFillColor (sf::Color (fill_color.r, fill_color.g, fill_color.b, fill_color.a));
+    rectangle.setOutlineThickness (1);
+    rectangle.setOutlineColor (sf::Color (outline_color.r, outline_color.g, outline_color.b, outline_color.a));
+    target.Draw (rectangle);
 }
 
 // ------------------------------------------------------------
 // Lines
 // ------------------------------------------------------------
-void Primitives::Lines(Window &target, const std::vector<Line> &lines, int thickness) {
+void Primitives::Lines (Window &target, const std::vector<Line> &lines, unsigned int thickness) {
 
     // construct the vertex list
     std::vector<sf::Vertex> vertices;
@@ -70,20 +70,22 @@ void Primitives::Lines(Window &target, const std::vector<Line> &lines, int thick
 
         // add the vertices
         for (auto it = lines.begin(); it != lines.end(); ++it) {
-            vertices.push_back(sf::Vertex(sf::Vector2f((*it).x1 + i * 0.2f, (*it).y1 + i * 0.2f), sf::Color(outline_color.r, outline_color.g, outline_color.b, outline_color.a)));
-            vertices.push_back(sf::Vertex(sf::Vector2f((*it).x2 + i * 0.2f, (*it).y2 + i * 0.2f), sf::Color(outline_color.r, outline_color.g, outline_color.b, outline_color.a)));
+            vertices.push_back (
+                sf::Vertex (sf::Vector2f ( (*it).x1 + i * 0.2f, (*it).y1 + i * 0.2f), sf::Color (outline_color.r, outline_color.g, outline_color.b, outline_color.a)));
+            vertices.push_back (
+                sf::Vertex (sf::Vector2f ( (*it).x2 + i * 0.2f, (*it).y2 + i * 0.2f), sf::Color (outline_color.r, outline_color.g, outline_color.b, outline_color.a)));
         }
     }
 
     // draw vertex list
-    target.window.draw(&vertices[0], vertices.size(), sf::Lines);
+    target.window.draw (&vertices[0], vertices.size(), sf::Lines);
 }
 
-void Primitives::DrawLine(Window &target, const Vector3 &point1, const Vector3 &point2, int thickness) {
+void Primitives::DrawLine (Window &target, const Vector3 &point1, const Vector3 &point2, unsigned int thickness) {
     sf::Vertex vertices[4];
     Vector3 direction = point2 - point1;
-    Vector3 unit_direction = direction / std::sqrt(direction.x * direction.x + direction.y * direction.y);
-    Vector3 unit_perpendicular(-unit_direction.y, unit_direction.x);
+    Vector3 unit_direction = direction / std::sqrt (direction.x * direction.x + direction.y * direction.y);
+    Vector3 unit_perpendicular (-unit_direction.y, unit_direction.x);
 
     Vector3 offset = unit_perpendicular * (thickness / 2.f);
 
@@ -105,40 +107,40 @@ void Primitives::DrawLine(Window &target, const Vector3 &point1, const Vector3 &
     vertices[3].position.y = tmp4.y;
 
     for (int i = 0; i < 4; ++i) {
-        vertices[i].color = sf::Color(outline_color.r, outline_color.g, outline_color.b, outline_color.a);
+        vertices[i].color = sf::Color (outline_color.r, outline_color.g, outline_color.b, outline_color.a);
     }
 
-    target.window.draw(vertices, 4, sf::Quads);
+    target.window.draw (vertices, 4, sf::Quads);
 }
 
 // ------------------------------------------------------------
 // Arc
 // ------------------------------------------------------------
-void Primitives::Arc(Window &target, double x, double y, double radius, double start, double end, int segments, int thickness) {
+void Primitives::Arc (Window &target, float x, float y, float radius, float start, float end, unsigned int segments, unsigned int thickness) {
 
     // construct the vertex list
     std::vector<sf::Vertex> vertices;
 
     // trackers
-    double act_start = start;
-    double act_end = end;
+    float act_start = start;
+    float act_end = end;
 
-    double lastx = 0;
-    double lasty = 0;
+    float lastx = 0;
+    float lasty = 0;
 
     for (unsigned int i = 0; i < thickness; ++i) {
-        double current = act_start;
+        float current = act_start;
         int segments_drawn = 0;
 
         while (current <= act_end) {
-            double ax = x + radius * cos(RADIANS(current));
-            double ay = y + radius * sin(RADIANS(current));
-            current += ((act_end - act_start) / segments);
+            float ax = x + radius * cosf (RADIANS (current));
+            float ay = y + radius * sinf (RADIANS (current));
+            current += ( (act_end - act_start) / segments);
 
-            vertices.push_back(sf::Vertex(sf::Vector2f(ax, ay), sf::Color(outline_color.r, outline_color.g, outline_color.b, outline_color.a)));
+            vertices.push_back (sf::Vertex (sf::Vector2f (ax, ay), sf::Color (outline_color.r, outline_color.g, outline_color.b, outline_color.a)));
 
             if (segments_drawn) {
-                DrawLine(target, Vector3(lastx, lasty), Vector3(ax, ay), thickness);
+                DrawLine (target, Vector3 (lastx, lasty), Vector3 (ax, ay), thickness);
                 // target.Present();
             }
 
@@ -157,12 +159,12 @@ void Primitives::Arc(Window &target, double x, double y, double radius, double s
     // target.window.draw(&vertices[0], vertices.size(), sf::LinesStrip);
 }
 
-void Primitives::OutlineColor(const Color &c) {
+void Primitives::OutlineColor (const Color &c) {
     last_outline_color = outline_color;
     outline_color = c;
 }
 
-void Primitives::FillColor(const Color &c) {
+void Primitives::FillColor (const Color &c) {
     last_fill_color = fill_color;
     fill_color = c;
 }
