+++ src/graphics/sprite.cpp
@@ -12,15 +12,15 @@ Sprite::Sprite(const std::string &filename, unsigned int num_rows, unsigned int 
   sprite_height = texture->getSize().y / num_rows;
 
   // set up the coordinates of each frame
-  for (int row = 0; row < num_rows; row++) {
-    for (int col = 0; col < num_cols; col++) {
+  for (unsigned int row = 0; row < num_rows; row++) {
+    for (unsigned int col = 0; col < num_cols; col++) {
 
       // sprite rectangle
       sf::IntRect rect;
-      rect.left = col * sprite_width;
-      rect.top = row * sprite_height;
-      rect.width = sprite_width;
-      rect.height = sprite_height;
+      rect.left = int(col * sprite_width);
+      rect.top = int(row * sprite_height);
+      rect.width = int(sprite_width);
+      rect.height = int(sprite_height);
       rects.push_back(rect);
     }
   }
@@ -47,7 +47,7 @@ void Sprite::SetFrame(unsigned int frame) {
 // ------------------------------------------------------------------
 // AddAnimation
 // ------------------------------------------------------------------
-void Sprite::AddAnimation(const std::string &animname, double frametime, bool loopanim, std::vector<int> framelist) {
+void Sprite::AddAnimation(const std::string &animname, float frametime, bool loopanim, std::vector<unsigned int> framelist) {
   AddAnimation(Animation(animname, frametime, loopanim, framelist));
 }
 
@@ -83,7 +83,7 @@ void Sprite::Animate() {
 // ------------------------------------------------------------------
 // Move
 // ------------------------------------------------------------------
-void Sprite::Move(double dx, double dy) {
+void Sprite::Move(float dx, float dy) {
   Renderable::Move(dx, dy);
   sprite.setPosition(geometry.x, geometry.y);
 }
@@ -111,21 +111,21 @@ void Sprite::SetAnimation(const std::string &name) {
 // ------------------------------------------------------------------
 // Scale
 // ------------------------------------------------------------------
-void Sprite::Scale(double x_factor, double y_factor) {
+void Sprite::Scale(float x_factor, float y_factor) {
   sprite.setScale(x_factor, y_factor);
 }
 
 // ------------------------------------------------------------------
 // Scale
 // ------------------------------------------------------------------
-void Sprite::Scale(double factor) {
+void Sprite::Scale(float factor) {
   sprite.setScale(factor, factor);
 }
 
 // ------------------------------------------------------------------
 // SetAnimationSpeed
 // ------------------------------------------------------------------
-void Sprite::SetAnimationSpeed(int time) {
+void Sprite::SetAnimationSpeed(unsigned int time) {
   if (current_animation) {
     current_animation->SetSpeed(time);
   }
