+++ src/statemachine/state.h
@@ -7,7 +7,7 @@ namespace GameLib {
 
 /** @brief base state */
 class State {
-  public:
+public:
     /**
      * @brief constructor
      */
@@ -27,7 +27,7 @@ class State {
      * @brief onStep
      * @param _dt time delta
      */
-    virtual void OnStep(const double _dt) = 0;
+    virtual void OnStep (const float _dt) = 0;
 
     /**
      * @brief onEnd
@@ -51,7 +51,7 @@ class State {
         return name;
     }
 
-  protected:
+protected:
     // identifier
     std::string name;
 };
